refactor(groups): calculate group color in core business logic
Additionally remove unnecessary real group and orpan logic that was previously
used to separate types of groups based on count for simpler color indexing but
is no longer as relevant due to changes made previously to handle windows.
